requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
datasets>=2.15.0
accelerate>=0.27.0
evaluate>=0.4.0
pyyaml>=6.0.0
pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.23.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
datasets>=2.16.0
accelerate>=0.27.0
evaluate>=0.4.0
Pillow>=10.0.0
pyyaml>=6.0.0
numpy>=1.24.0
tqdm>=4.66.0

